+++ src/utils/helpers.js
@@ -30,19 +30,24 @@ export function indexOf(array, obj, fn) {
  */
 export const isMobile = {
     Android: function () {
-        return navigator.userAgent.match(/Android/i)
+        return typeof window !== 'undefined' &&
+            window.navigator.userAgent.match(/Android/i)
     },
     BlackBerry: function () {
-        return navigator.userAgent.match(/BlackBerry/i)
+        return typeof window !== 'undefined' &&
+            window.navigator.userAgent.match(/BlackBerry/i)
     },
     iOS: function () {
-        return navigator.userAgent.match(/iPhone|iPad|iPod/i)
+        return typeof window !== 'undefined' &&
+            window.navigator.userAgent.match(/iPhone|iPad|iPod/i)
     },
     Opera: function () {
-        return navigator.userAgent.match(/Opera Mini/i)
+        return typeof window !== 'undefined' &&
+            window.navigator.userAgent.match(/Opera Mini/i)
     },
     Windows: function () {
-        return navigator.userAgent.match(/IEMobile/i)
+        return typeof window !== 'undefined' &&
+            window.navigator.userAgent.match(/IEMobile/i)
     },
     any: function () {
         return (isMobile.Android() || isMobile.BlackBerry() || isMobile.iOS() || isMobile.Opera() || isMobile.Windows())